+++ changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ## 97bc4346faf5cfe7c3e87cdba0a31da248221d2d - 2017-11-29 - Yudhvir
 ### Added
+- Map icon opens map fragment
+
+## 97bc4346faf5cfe7c3e87cdba0a31da248221d2d - 2017-11-29 - Yudhvir
+### Added
 - Push notifications work and display on ui thread
 
 ## 97bc4346faf5cfe7c3e87cdba0a31da248221d2d - 2017-11-29 - Yudhvir